refactor: gradient && support more degreee unit
